requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
jinja2==3.1.3
python-multipart==0.0.6
reportlab==4.0.9
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
Pillow==10.2.0
